Began collecting refreshed results database
Initiated a new project, rebuilding the results database with verified
citations. As long as I'm at it, I'm organizing by paper and chronology
(at least to year of publication).

NB: This version is incomplete and WILL NOT compile as a results
database. The current byPaper.txt file is not yet completely consistent
in its notation. Also, some papers mentioned are yet to be indexed,
others are still absent, and yet more are currently omitted entirely, as
I haven't gotten to them yet.