Cleaned up the new results database
Beginning to standardize on the new annotated-bibliography version of
the results database; now runs smoothly, and confirms almost all
knowledge present in the previous database.

Outstanding issues:
1) It is not currently possible to query a derivable fact about a
conjunction not mentioned in the collection of known results.
2) Some results in the previous database (mostly cited as "Bienvenu,
Patey, and Shafer (to appear)", as well as BSig2+COH+RRT22 -|> ISig2
"Wang-TA") cannot be confirmed from external resources.