Switched to bottom-up complexity optimization
For better performance (~20% faster), switched rmupdater from a cached
top-down computation of complexity of justifications to a bottom-up
computation.

As a side effect, I was forced to remove support for the "-f" option;
the only other option that maintained the performance advantage would
have been to add a second logic path for EVERY inference rule, with a
great deal of repeated code.