Fixed support for Python 2.7
Made the necessary changes to codebase for simultaneous support for
Python 2 and 3; provided with the enum34 backport, this system should
function on any version of Python at 2.7 or higher.

Now that it's possible, it seems the RM Zoo runs about 33% faster on
PyPy (with lru_cache disabled), at the cost of some lowered performance
(and increased space usage?) for pickle. However, we pay a significant
performance penalty on CPython 2.7; I will try to improve this situation
soon.